Fixed Warnings, added some things
Added the functions Unk00, Unk01 from the CdSys class.
They are not executed because of the issues with library linking for some reason (Refer to line 75)

Fixed the warnings for HWD0 & VWD0.
Fixed the warning for Clip3F, it wass expecting an EVECTOR **, butt passing &evmx, except that evmx is already a pointer.